requirements.txt
Flask
python-dotenv
eventlet
gunicorn
loguru
requests
ipython
# mediapipe
opencv-python-headless
tqdm
handyllm
pyvis
pyyaml